minor fixes

+++ include/my_funcs.h
@@ -8,7 +8,7 @@ int comparaDatas(Date a, Date b);
 int ordena(gconstpointer a,gconstpointer b);
 void printUserHT(gpointer key, gpointer value, gpointer user_data);
 void printPostHTa(gpointer key, gpointer value, gpointer user_data);
-int insert(long* array, long elem, int size);
+int insert(int* array, int elem, int size);
 int insertDate(Date* array, Date elem, int size);
 void insereId(long* v, long x, int i, int n);
 char* envolveTag(char* tag);
